🚚 move: RankDefault & RankUser 분리
