--- app/src/components/elements/DashboardContentView/Rank/index.ts
@@ -1,2 +1,0 @@
-export * from './Rank';
-export * from './RankUser';
