+++ app/src/components/elements/DashboardContentView/RankDefault/index.tsx
@@ -2,13 +2,13 @@ import { VStack } from '@components/common';
 import type { RankItemType } from '@utils/types/Rank';
 import { RankItem } from './RankItem';
 
-type RankProps = {
+type RankDefaultProps = {
   rankList: RankItemType[];
   cnt: number;
   unit: string;
 };
 
-export const Rank = ({ rankList, cnt, unit }: RankProps) => {
+export const RankDefault = ({ rankList, cnt, unit }: RankDefaultProps) => {
   return (
     <VStack w="100%" h="100%" spacing="2rem">
       {rankList.slice(0, cnt).map((rankItem) => (
